src/mcp_servers/blip-mediphi-bnb/requirements.txt
mcp
accelerate
transformers
sentencepiece
Pillow
huggingface_hub
pydantic
bitsandbytes
hydra-core
flash-attn
protobuf